requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
loguru==0.7.2
chromadb==0.4.22
sentence-transformers==2.5.1
google-generativeai==0.3.2
python-dotenv==1.0.1
pyyaml==6.0.1